% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_pdqr_funs}
\alias{add_pdqr_funs}
\title{Add columns for pdqr functions (i.e., functions to evaluate the cdf, pdf, or
quantile function, or to generate random deviates) determined by parameter columns}
\usage{
add_pdqr_funs(
  df,
  dist = df$dist,
  trans = NULL,
  trans_inv = NULL,
  transpars = NULL,
  types = c("p", "d", "q", "r"),
  fnames = c(p = "F", d = "f", q = "Q", r = "r")
)
}
\arguments{
\item{df}{data frame; must have columns for each parameter required by any
of the distribution functions indicated in a column named \code{dist} provided directly
or via argument \code{dist}}

\item{dist}{character vector of length 1 or \code{nrow(df)} containing distribution
root names (or \code{distfromq}); must be provided if df does not have a column \code{dist}}

\item{trans}{function with arguments x for outcome and parameters for a transformation
precomposed with pd functions.}

\item{trans_inv}{inverse of trans with argument p for probability level; will be
post-composed with qr function.}

\item{transpars}{parameters for trans and trans_inv not included in columns of df}

\item{types}{which pdqr function list columns to add}

\item{fnames}{names of pdqr function list columns}
}
\value{
a tibble with list columns containing selected pdqr functions
}
\description{
Add columns for pdqr functions (i.e., functions to evaluate the cdf, pdf, or
quantile function, or to generate random deviates) determined by parameter columns
}
