% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{alloscore.slim}
\alias{alloscore.slim}
\title{Score a slim allocated data frame}
\usage{
\method{alloscore}{slim}(slim_df, ys, against_oracle)
}
\arguments{
\item{ys}{a list of named outcome vectors with names matching target names in slim_df}
}
\value{
a tibble with columns \code{xdf} and \code{scores} containing the components and scores for
each sample in \code{ys}
}
\description{
Score a slim allocated data frame
}
