% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{dexp_gpl_df}
\alias{dexp_gpl_df}
\title{Create a list of derivatives of expected gpl losses from a gpl_df}
\usage{
dexp_gpl_df(df, F)
}
\arguments{
\item{F}{list of predictive cdfs}
}
\description{
Create a list of derivatives of expected gpl losses from a gpl_df
}
