% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{dexp_gpl_loss}
\alias{dexp_gpl_loss}
\title{Create derivative of expected gpl loss}
\usage{
dexp_gpl_loss(g = "x", dg = NULL, F, kappa = 1, alpha = NA, O, U = NA)
}
\arguments{
\item{U}{}
}
\value{
function with argument x giving the derivative of expected loss with respect to the
distribution F
}
\description{
Create derivative of expected gpl loss
}
