% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{dexp_over_loss}
\alias{dexp_over_loss}
\title{Create derivative of expected over-prediction loss}
\usage{
dexp_over_loss(g = "x", dg = NULL, F)
}
\arguments{
\item{dg}{}
}
\description{
Create derivative of expected over-prediction loss
}
