% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{exp_under_loss}
\alias{exp_under_loss}
\title{Basic expected g-linear loss for under-prediction of a random outcome Y}
\usage{
exp_under_loss(
  dg = function(y) {
     1
 },
  F
)
}
\arguments{
\item{dg}{derivative of g, probably will work even with discontinuities}

\item{F}{predictive cdf of an outcome Y}
}
\value{
function of prediction x giving expected g-linear loss for under-predicting Y~F
}
\description{
Basic expected g-linear loss for under-prediction of a random outcome Y
}
