% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{expected_over_loss}
\alias{expected_over_loss}
\title{Basic expected g-linear loss for over-prediction of a random outcome Y}
\usage{
expected_over_loss(
  dg = function(y) {
     1
 },
  F
)
}
\arguments{
\item{dg}{derivative of g, probably will work even with discontinuities}

\item{F}{predictive cdf of an outcome Y}
}
\value{
function of prediction x giving expected g-linear loss for over-predicting Y~F
}
\description{
Basic expected g-linear loss for over-prediction of a random outcome Y
}
