% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_args_from_df}
\alias{get_args_from_df}
\title{Utility function to make data frames of easier to work with}
\usage{
get_args_from_df(df)
}
\description{
Utility function to make data frames of easier to work with
}
