% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_derivative}
\alias{get_derivative}
\title{Get the derivative of a function of x entered as a string, e.g., "log(x)"}
\usage{
get_derivative(func)
}
\description{
Get the derivative of a function of x entered as a string, e.g., "log(x)"
}
