% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{gpl}
\alias{gpl}
\title{Get the \code{gpl_df} attribute of an allocated data frame which containes gpl loss functions
for each target and the parameters used to form them.}
\usage{
gpl(adf)
}
\arguments{
\item{adf}{an allocated data frame}
}
\value{
data frame
}
\description{
Get the \code{gpl_df} attribute of an allocated data frame which containes gpl loss functions
for each target and the parameters used to form them.
}
