% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{gpl_loss_exp_fun}
\alias{gpl_loss_exp_fun}
\title{Create gpl expected loss function}
\usage{
gpl_loss_exp_fun(..., gpl_loss = NULL, f)
}
\arguments{
\item{...}{passed to gpl_loss_fun}

\item{gpl_loss}{specified loss function if ... missing}

\item{f}{probability density}
}
\value{
function with argument x giving the expected loss with respect to the
distribution f
}
\description{
Create gpl expected loss function
}
