% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{make_slim}
\alias{make_slim}
\title{Make a slim allocated data frame with columns only for K, target_name, \code{x},
and optionally scoring functions with \code{x} inserted to give functions of observed \code{y}.}
\usage{
make_slim(adf, keep_score_fun = TRUE)
}
\arguments{
\item{adf}{an allocated data frame}

\item{keep_score_fun}{}
}
\description{
Make a slim allocated data frame with columns only for K, target_name, \code{x},
and optionally scoring functions with \code{x} inserted to give functions of observed \code{y}.
}
