% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{margexb_fun}
\alias{margexb_fun}
\title{/ate function to calculate the marginal expected benefit of allocating an
additional unit of resources to a given target.}
\usage{
margexb_fun(F, kappa = 1, alpha, w, g = NULL, dg = 1, q_scale = 1, ...)
}
\arguments{
\item{F}{cdf for forecast distribution}

\item{kappa}{scale factor}

\item{alpha}{normalized loss when outcome y exceeds forecast x}

\item{w}{weight for given target}

\item{g}{string describing increment function which will be used
to form gpl function for this target; should be written using \code{x}, e.g. "log(x)";
will be differentiated using \code{stats::D} so don't try anything fancy;
defaults to identity via default of 1 for dg.}

\item{dg}{explicit numeric constant or function giving the derivative of \code{g} when \code{g}
is not specified. Defaults to 1 corresponding to pinball loss.}

\item{q_scale}{scaling factor for resource level x; used for aligning marginal expected
benefit functions of different targets (and parameters) on single x interval.}

\item{...}{ignored}
}
\value{
a function with argument x that calculates the marginal expected benefit
}
\description{
/ate function to calculate the marginal expected benefit of allocating an
additional unit of resources to a given target.
}
\details{
Note that this is minus the derivative of the expected score \eqn{\overline{s}_F}
}
