% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{meb_gpl_df}
\alias{meb_gpl_df}
\title{Create marginal expected benefit functions for vector of gpl losses with respect
to predictive distributions F and weights w}
\usage{
meb_gpl_df(df, F, w)
}
\arguments{
\item{F}{list of predictive cdfs}

\item{w}{weights for optimizing allocation}
}
\description{
Create marginal expected benefit functions for vector of gpl losses with respect
to predictive distributions F and weights w
}
