% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{new_gpl_df}
\alias{new_gpl_df}
\title{Create data frame of gpl functions and associated parameters}
\usage{
new_gpl_df(
  N = NULL,
  target_names = NA,
  g = "x",
  kappa = 1,
  alpha = 1,
  O = NA,
  U = NA,
  offset = 0
)
}
\arguments{
\item{offset}{}
}
\description{
Create data frame of gpl functions and associated parameters
}
