% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{normalize_col_names}
\alias{normalize_col_names}
\title{Column name normalizing utility}
\usage{
normalize_col_names(
  slim_df,
  origin_time_col_name,
  model_col_name,
  target_col_name
)
}
\arguments{
\item{target_col_name}{}
}
\description{
Column name normalizing utility
}
