% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{oracle_allocate}
\alias{oracle_allocate}
\title{Allocate according to an oracle's knowledge of outcome y}
\usage{
oracle_allocate(gpl_df, y, K, w, ...)
}
\arguments{
\item{gpl_df}{a \code{gpl_df} object}

\item{y}{observed outcomes}

\item{K}{vector of constraints on total provision (cannot be supplied via df)}

\item{w}{numeric vector with cost per unit resource allocated to each
coordinate}
}
\value{
see \code{allocate}
}
\description{
Allocate according to an oracle's knowledge of outcome y
}
