% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{oracle_allocate_direct}
\alias{oracle_allocate_direct}
\title{Directly find the oracle allocation with no data management.
Assumes for now that all targets have same gpl parameters and
(probably?) that \code{g = x}}
\usage{
oracle_allocate_direct(y, K, w)
}
\arguments{
\item{w}{}
}
\description{
Directly find the oracle allocation with no data management.
Assumes for now that all targets have same gpl parameters and
(probably?) that \code{g = x}
}
