% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{oracle_alloscore}
\alias{oracle_alloscore}
\title{Score according to an oracle's knowledge of outcome y}
\usage{
oracle_alloscore(df = NULL, y, K, kappa = 1, alpha = 1, g = function(u) u, ...)
}
\arguments{
\item{df}{data frame containing columns for other list arguments which are used only when
those arguments (e.g. \code{F}) are not passed directly}

\item{y}{numeric observed data value}

\item{K}{vector of constraints on total provision}

\item{kappa}{scale factor}

\item{alpha}{normalized loss when outcome y exceeds forecast x}

\item{g}{list of functions that calculate the gpl function for each coordinate;
default value of NULL causes pinball loss to be used}
}
\value{
see \code{alloscore}
}
\description{
Score according to an oracle's knowledge of outcome y
}
