% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{oracle_alloscore_direct}
\alias{oracle_alloscore_direct}
\title{Directly find and score the oracle allocation with no data management.}
\usage{
oracle_alloscore_direct(y, K, w, gpl_fns)
}
\arguments{
\item{gpl_fns}{}
}
\description{
Directly find and score the oracle allocation with no data management.
}
