% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{over_loss}
\alias{over_loss}
\title{Basic g-linear loss for over-prediction of an outcome y}
\usage{
over_loss(g = function(u) {
    u
})
}
\arguments{
\item{g}{a non-decreasing function}
}
\value{
function of prediction x and outcome y that only penalizes over-prediction
}
\description{
Basic g-linear loss for over-prediction of an outcome y
}
