% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pdqr_factory}
\alias{pdqr_factory}
\title{Produce a pdqr function with set parameters}
\usage{
pdqr_factory(..., dist, type, trans = NULL, trans_inv = NULL, transpars = NULL)
}
\arguments{
\item{...}{list of parameter name-value pairs to be partialed out}

\item{dist}{distribution root name or \code{distfromq}}

\item{type}{p, d, q, or r}

\item{trans}{function with arguments x for outcome and parameters for a transformation
precomposed with pd functions.}

\item{trans_inv}{inverse of trans with argument p for probability level; will be
post-composed with qr function.}

\item{transpars}{parameters for trans and trans_inv not included in columns of df}
}
\value{
function of type p, d, q, or r
}
\description{
Produce a pdqr function with set parameters
}
