% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_components.allocated}
\alias{plot_components.allocated}
\title{Default plot_components method; used to plot multiple allocated dfs.}
\usage{
\method{plot_components}{allocated}(
  ...,
  Ks = NULL,
  origin_time_col_name = NULL,
  model_col_name = NULL,
  target_col_name = NULL
)
}
\arguments{
\item{...}{allocated data frames}

\item{Ks}{Ks to plot over, defaults to NULL which keeps all. Note that some plots
require a single K}

\item{origin_time_col_name}{name of of origin time columns}

\item{model_col_name}{name of model columns}

\item{target_col_name}{name of columns containing target names}
}
\description{
Default plot_components method; used to plot multiple allocated dfs.
}
