% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_components.default}
\alias{plot_components.default}
\title{Plot components of a nested data frame of scored allocations}
\usage{
\method{plot_components}{default}(
  df,
  Ks = NULL,
  scored_col_name = "scored",
  origin_time_col_name = NULL,
  model_col_name = NULL,
  target_col_name = NULL
)
}
\arguments{
\item{target_col_name}{}
}
\description{
Plot components of a nested data frame of scored allocations
}
