% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_components.slim}
\alias{plot_components.slim}
\title{Plot components}
\usage{
\method{plot_components}{slim}(
  slim_df,
  Ks = NULL,
  origin_time_col_name = NULL,
  model_col_name = NULL,
  target_col_name = NULL
)
}
\arguments{
\item{target_col_name}{}
}
\description{
Plot components
}
