% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_components_slim}
\alias{plot_components_slim}
\title{Plot components}
\usage{
plot_components_slim(
  slim_df,
  Ks = NULL,
  origin_time_col_name = NULL,
  model_col_name = NULL,
  target_col_name = NULL,
  show_oracle = TRUE,
  show_raw = TRUE,
  order_at_K = NULL,
  order_at_model = NULL,
  order_at_origin_time = NULL,
  pal_top = c("#CD3333", "#009ACD", "#FFB90F", "#8B3E2F", "#8B8B00"),
  bar_positioning = "stack"
)
}
\arguments{
\item{target_col_name}{}
}
\description{
Plot components
}
