% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_iterations}
\alias{plot_iterations}
\title{Plot allocation iteration for a particular K}
\usage{
plot_iterations(
  adf,
  K_to_plot = NULL,
  itnum = NULL,
  num_targets_to_color = 6,
  target_palette = (scales::viridis_pal())(num_targets_to_color + 1)
)
}
\arguments{
\item{adf}{an allocate data frame}

\item{K_to_plot}{the constraint level to show iterations for}

\item{itnum}{how many iterations to show; default NULL}

\item{target_palette}{}
}
\description{
Plot allocation iteration for a particular K
}
