% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_scores_slim}
\alias{plot_scores_slim}
\title{Plot allocation scores}
\usage{
plot_scores_slim(
  slim_df,
  Ks = NULL,
  ts = FALSE,
  ts_dates = NULL,
  origin_time_col_name = "reference_date",
  model_col_name = "model",
  target_col_name = "abbreviations",
  show_oracle = FALSE,
  show_raw = FALSE,
  order_at_K = NULL,
  order_at_model = NULL,
  order_at_origin_time = NULL,
  palette,
  linetypes,
  ytot = TRUE
)
}
\arguments{
\item{target_col_name}{}
}
\description{
Plot allocation scores
}
