% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{post_process}
\alias{post_process}
\title{Post process allocations that are infeasible due to plateaus}
\usage{
post_process(x, K, lam, w, Lambda, eps_lam, point_mass_window)
}
\arguments{
\item{x}{last iteration of allocation via lambda search}

\item{K}{constraint}

\item{lam}{last lambda iterate}

\item{w}{weights}

\item{Lambda}{meb's}

\item{eps_lam}{lambda tolerance that led to lambda convergence failure}

\item{point_mass_window}{}
}
\description{
Post process allocations that are infeasible due to plateaus
}
