% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{slim}
\alias{slim}
\title{Make an allocated data frame slim, i.e., containing
columns only for scores and optionally a list column or unnested
columns for the \code{xdf} component data frames of the allocation}
\usage{
slim(adf, xdf_action = c("default", "unnest", "nest"), id_cols = NULL)
}
\arguments{
\item{adf}{an allocated data frame}

\item{xdf_action}{character indicating whether to unnest \code{xdf}.
Default is to unnest when data from is scored and leave nested
when data frame is scored.}

\item{id_cols}{columns to also keep such as model name and origin time}
}
\description{
Make an allocated data frame slim, i.e., containing
columns only for scores and optionally a list column or unnested
columns for the \code{xdf} component data frames of the allocation
}
