% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{stdize_news_params}
\alias{stdize_news_params}
\title{Convert newsvendor parameters to kappa, alpha}
\usage{
stdize_news_params(ax, ay = NULL, a_minus, a_plus)
}
\arguments{
\item{ax}{wholesale cost}

\item{a_plus}{}
}
\value{
kappa and alpha as list
}
\description{
Convert newsvendor parameters to kappa, alpha
}
