% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stdize_ou_params}
\alias{stdize_ou_params}
\title{Convert over/underprediction parameters to kappa, alpha}
\usage{
stdize_ou_params(O, U)
}
\arguments{
\item{O}{incremental loss that is incurred when overprediction leads to unused supply}

\item{U}{incremental loss that is incurred when underprediction leads to unmet need}
}
\value{
kappa and alpha as list
}
\description{
Convert over/underprediction parameters to kappa, alpha
}
