% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{under_loss}
\alias{under_loss}
\title{Basic g-linear loss for under-prediction of an outcome y}
\usage{
under_loss(
  g = function(u) {
     u
 }
)
}
\arguments{
\item{g}{a non-decreasing function}
}
\value{
function of prediction x and outcome y that only penalizes under-prediction
}
\description{
Basic g-linear loss for under-prediction of an outcome y
}
