% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unirootL}
\alias{unirootL}
\title{Find an approximate root for a decreasing function to the left of the root}
\usage{
unirootL(f, lower, upper, point_mass_window)
}
\arguments{
\item{f}{the function for which the root is sought.}

\item{lower, upper}{the lower and upper end points of the interval to
    be searched.}

\item{point_mass_window}{a distance to move to the left in order to find approximate root
with non-negative value}
}
\value{
Either root element of \code{stats::uniroot} evaluated on \code{f} or something slightly
to its left.
}
\description{
Find an approximate root for a decreasing function to the left of the root
}
