% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{weights.allocated}
\alias{weights.allocated}
\title{Get the weights of an allocated data frame}
\usage{
\method{weights}{allocated}(adf)
}
\arguments{
\item{adf}{an allocated data frame}
}
\description{
Get the weights of an allocated data frame
}
