% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zxh_tab2}
\alias{zxh_tab2}
\title{Data from Table 2 in ZXH for the normal distribution example}
\format{
An object of class \code{data.frame} with 17 rows and 9 columns.
}
\source{
B. Zhang, X. Xu, and Z. Hua, “A binary solution method for the multiproduct newsboy problem with budget constraint,”
Int. J. Prod. Econ., vol. 117, no. 1, pp. 136–141, 2009.
}
\usage{
zxh_tab2
}
\description{
A data frame with 17 rows and 9 columns:
\tabular{ll}{
\strong{Product} \tab	Product number \cr
\strong{v}		 \tab	cost of revenue loss per unit not stocked of product \emph{i}\cr
\strong{h}		 \tab	cost incurred per unit leftover of product \emph{i}	\cr
\strong{c}		 \tab	cost per unit of product \emph{i}\cr
\strong{mu}		 \tab	mean of demand for product \emph{i}\cr
\strong{sigma}	 \tab	standard deviation of demand for product \emph{i}\cr
\strong{q_zxh}	 \tab	\emph{(v-c)/(h+v)} quantile of demand distribution as given by ZXH\cr
\strong{GIM}	 \tab	solution of Abdel-Malek and Montanari (2005a)\cr
\strong{Opt}	 \tab	solution of ZXH
}
}
\keyword{datasets}
