% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zxh_tab3}
\alias{zxh_tab3}
\title{Data from Table 3 in ZXH for the beta distribution example}
\format{
An object of class \code{data.frame} with 6 rows and 11 columns.
}
\source{
B. Zhang, X. Xu, and Z. Hua, “A binary solution method for the multiproduct newsboy problem with budget constraint,”
Int. J. Prod. Econ., vol. 117, no. 1, pp. 136–141, 2009.
}
\usage{
zxh_tab3
}
\description{
A data frame with 6 rows and 11 columns:
\tabular{ll}{
\strong{Product} \tab	Product number \cr
\strong{v}		 \tab	cost of revenue loss per unit not stocked of product \emph{i}\cr
\strong{h}		 \tab	cost incurred per unit leftover of product \emph{i}	\cr
\strong{c}		 \tab	cost per unit of product \emph{i}\cr
\strong{x_min}	 \tab	min of demand distribution support\cr
\strong{x_max}	 \tab	max of demand distribution support\cr
\strong{Balpha}	 \tab	alpha (i.e. shape1) parameter of demand distribution\cr
\strong{Bbeta}	 \tab	beta (i.e. shape2) parameter of demand distribution\cr
\strong{GIM}	 \tab	solution of Abdel-Malek and Montanari (2005a)\cr
\strong{Opt}	 \tab	solution of ZXH
}
}
\keyword{datasets}
