% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{allocate}
\alias{allocate}
\title{Allocate to minimize expected gpl loss under forecasts F with constraint K}
\usage{
allocate(
  df = NULL,
  F,
  Q,
  w = 1,
  K,
  kappa = 1,
  alpha = 1,
  dg = 1,
  eps_K = 0.01,
  eps_lam = 1e-05,
  point_mass_window = 0.001,
  Trace = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{df}{data frame containing columns for other list arguments which are used only when
those arguments (e.g. \code{F}) are not passed directly}

\item{F}{list of cdf functions for forecast distributions}

\item{Q}{list of quantile functions for forecast distributions}

\item{w}{numeric vector with cost per unit resource allocated to each
coordinate}

\item{K}{constraint on total provision}

\item{kappa}{scale factor}

\item{alpha}{normalized loss when outcome y exceeds forecast x}

\item{dg}{numeric constant(s) or function(s) to calculate the derivative of the
gpl function \code{g} for each coordinate}

\item{Trace}{logical, record iterations}

\item{Verbose}{logical, echo K and report number of iterations required to find allocation}
}
\value{
If \code{Trace} is \code{FALSE}, a numeric vector of length \code{N} givig the
optimal allocations. If \code{Trace} is \code{TRUE}, a named list with entries:
\itemize{
\item \code{x}: numeric vector of length \code{N} giving the optimal allocations
\item \code{xs}: list of values of \code{x} over the course of optimization
\item \code{lambdas}: numeric vector of values of lambda over the course of optimization
}
}
\description{
Allocate to minimize expected gpl loss under forecasts F with constraint K
}
