% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{allocate}
\alias{allocate}
\title{Allocate to minimize expected gpl loss under forecasts F with constraint K}
\usage{
allocate(
  df = NULL,
  F,
  Q,
  w = 1,
  K,
  kappa = 1,
  alpha = 1,
  g = NULL,
  dg = 1,
  eps_lam = 1e-04,
  eps_K = 0.01,
  point_mass_window = 0.001
)
}
\arguments{
\item{df}{data frame containing columns for other list arguments which are used only when
those arguments (e.g. \code{F}) are not passed directly}

\item{F}{list of cdf functions for forecast distributions}

\item{Q}{list of quantile functions for forecast distributions}

\item{w}{numeric vector with cost per unit resource allocated to each
coordinate}

\item{K}{vector of constraints on total provision}

\item{kappa}{scale factor}

\item{alpha}{normalized loss when outcome y exceeds forecast x}

\item{g}{list of strings describing increment functions which will be used
to form gpl functions for each coordinate; should be written using \code{x}, e.g. "log(x)";
will be differentiated using \code{stats::D} so don't try anything fancy; defaults to NULL
deferring to \code{dg}}

\item{dg}{numeric constant(s) or function(s) to calculate the derivative of the
increment function \code{g} for each coordinate; defaults to 1 corresponding to pinball loss}

\item{point_mass_window}{}
}
\value{
data frame with (list-)columns
\describe{
\item{K}{constraints}
\item{x}{allocations for constraints}
\item{xs}{array with allocations at each iteration as rows}
\item{lam_seq}{sequence of Lagrange multipliers tested at each iteration}
}
}
\description{
Allocate to minimize expected gpl loss under forecasts F with constraint K
}
