% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{allocate}
\alias{allocate}
\title{Allocate to minimize expected gpl loss under forecasts F with constraint K}
\usage{
allocate(F, Q, kappa, alpha, dg = 1, w, K, eps_K, eps_lam, Trace = FALSE)
}
\arguments{
\item{Trace}{logical, record iterations}
}
\value{

}
\description{
Allocate to minimize expected gpl loss under forecasts F with constraint K
}
