% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{alloscore}
\alias{alloscore}
\title{Obtain the allocation score for a given forecast distribution F for the
observed data value y in a constrained allocation problem.}
\usage{
alloscore(
  y,
  F,
  Q,
  w,
  K,
  kappa = 1,
  alpha,
  dg = 1,
  eps_K,
  eps_lam,
  g = NULL,
  against_oracle = TRUE
)
}
\arguments{
\item{y}{numeric observed data value}

\item{F}{list of cdf functions for forecast distributions}

\item{Q}{list of quantile functions for forecast distributions}

\item{w}{numeric vector with cost per unit resource allocated to each
coordinate}

\item{K}{constraint on total provision}

\item{kappa}{scale factor}

\item{alpha}{normalized loss when outcome y exceeds forecast x}

\item{dg}{numeric constant(s) or function(s) to calculate the derivative of the
gpl function \code{g} for each coordinate}

\item{g}{list of functions that calculate the gpl function for each coordinate;
default value of NULL causes pinball loss to be used}

\item{against_oracle}{logical; if \code{TRUE}, scores are normalized relative to
an oracle forecaster}
}
\description{
Obtain the allocation score for a given forecast distribution F for the
observed data value y in a constrained allocation problem.
}
