% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{alloscore}
\alias{alloscore}
\title{Obtain the allocation score for a given forecast distribution F for the
observed data value y in a constrained allocation problem.}
\usage{
alloscore(
  df = NULL,
  y,
  F,
  Q,
  w = 1,
  K,
  kappa = 1,
  alpha = 1,
  dg = 1,
  eps_K = 0.01,
  eps_lam = 1e-05,
  g = NULL,
  against_oracle = TRUE
)
}
\arguments{
\item{df}{data frame containing columns for other list arguments which are used only when
those arguments (e.g. \code{F}) are not passed directly}

\item{y}{numeric observed data value}

\item{F}{list of cdf functions for forecast distributions}

\item{Q}{list of quantile functions for forecast distributions}

\item{w}{numeric vector with cost per unit resource allocated to each
coordinate}

\item{K}{vector of constraints on total provision}

\item{kappa}{scale factor}

\item{alpha}{normalized loss when outcome y exceeds forecast x}

\item{dg}{numeric constant(s) or function(s) to calculate the derivative of the
gpl function \code{g} for each coordinate}

\item{g}{list of functions that calculate the gpl function for each coordinate;
default value of NULL causes pinball loss to be used}

\item{against_oracle}{logical; if \code{TRUE}, scores are normalized relative to
an oracle forecaster}

\item{components}{logical; if TRUE, the components of the score (relative to the components
of the oracle score if \code{against_oracle} is \code{TRUE}) are also returned}
}
\value{
a date frame of the form returned by \code{allocate} with additional columns for
\itemize{
\item \code{components_raw}: the raw gpl losses in each location
\item \code{score_raw}: the sum of the raw components
\item \code{components_oracle}: the gpl losses of an oracle in each location
\item \code{score_oracle}: the sum of the oracle's components
\item \code{components}: forecaster's difference from the oracle
\item \code{score}: the forecaster's score minus the oracle's
}
}
\description{
Obtain the allocation score for a given forecast distribution F for the
observed data value y in a constrained allocation problem.
}
