% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{alloscore}
\alias{alloscore}
\alias{alloscore.default}
\alias{alloscore.allocated}
\title{Score the allocations from set of forecasts against realized outcomes \code{y}.}
\usage{
alloscore(df = NULL, ...)

\method{alloscore}{default}(
  df = NULL,
  K,
  target_names = NA,
  y,
  F = NULL,
  Q = NULL,
  w = 1,
  kappa = 1,
  alpha = 1,
  g = "x",
  dg = NA,
  eps_K = 0.01,
  eps_lam = 1e-04,
  against_oracle = TRUE,
  slim = FALSE
)

\method{alloscore}{allocated}(df, y, against_oracle = TRUE)
}
\arguments{
\item{df}{an allocated data frame}

\item{K}{vector of constraints on total provision (cannot be supplied via df)}

\item{y}{numeric observed data value}

\item{F}{list of cdf functions for forecast distributions}

\item{Q}{list of quantile functions for forecast distributions}

\item{w}{numeric vector with cost per unit resource allocated to each
coordinate}

\item{kappa}{scale factor}

\item{alpha}{normalized loss when outcome y exceeds forecast x}

\item{g}{list of strings describing increment functions which will be used
to form gpl functions for each coordinate; should be written using \code{x}, e.g. "log(x)";
will be differentiated using \code{stats::D} so don't try anything fancy; defaults to NULL
deferring to \code{dg}}

\item{dg}{numeric constant(s) or function(s) to calculate the derivative of the
increment function \code{g} for each coordinate; defaults to 1 corresponding to pinball loss}

\item{against_oracle}{logical; if \code{TRUE}, components and scores relative to oracle are
included}

\item{slim}{logical; if TRUE use slim data frames}
}
\value{
a data frame of the form returned by \code{allocate} with additional columns for
\itemize{
\item \code{components_raw}: the raw gpl losses in each location
\item \code{score_raw}: the sum of the raw components
\item \code{components_oracle}: the gpl losses of an oracle in each location
\item \code{score_oracle}: the sum of the oracle's components
\item \code{components}: forecaster's difference from the oracle
\item \code{score}: the forecaster's score minus the oracle's
}
}
\description{
Score the allocations from set of forecasts against realized outcomes \code{y}.

Default method for alloscore, used when parameters passed individually; works by
allocating and then forwarding to allocated method.

Allocation scoring method for forecasts with computed allocations and gpl loss functions.
}
