% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{gpl_loss_exp_fun}
\alias{gpl_loss_exp_fun}
\title{Create gpl expected loss function}
\usage{
gpl_loss_exp_fun(
  dg = function(u) {
     1
 },
  F,
  kappa = 1,
  alpha,
  O,
  U,
  const = 0
)
}
\arguments{
\item{dg}{derivative of gpl function}

\item{F}{predictive CDF}

\item{kappa}{scale factor}

\item{alpha}{normalized loss when outcome y exceeds forecast x}

\item{O}{cost when forecast x exceeds outcome y; equals kappa*(1-alpha)}

\item{U}{loss when outcome y exceeds forecast x; equals kappa*alpha}
}
\value{
function with argument x giving the expected loss with respect to the
distribution F
}
\description{
Create gpl expected loss function
}
