% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{gpl_loss_fun}
\alias{gpl_loss_fun}
\title{Create gpl scoring/loss function}
\usage{
gpl_loss_fun(g = function(u) u, kappa = 1, alpha, O, U, const = 0)
}
\arguments{
\item{g}{gpl function}

\item{kappa}{scale factor}

\item{alpha}{normalized loss when outcome y exceeds forecast x}

\item{O}{cost when forecast x exceeds outcome y; equals kappa*(1-alpha)}

\item{U}{loss when outcome y exceeds forecast x; equals kappa*alpha}
}
\value{
function giving loss
}
\description{
Create gpl scoring/loss function
}
