% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{margexb_fun}
\alias{margexb_fun}
\title{Create function to calculate the marginal expected benefit of allocating an
additional unit of resources to a given target.}
\usage{
margexb_fun(F, kappa = 1, alpha, w, dg = 1, q_alpha = 1, ...)
}
\arguments{
\item{F}{list of cdf functions for forecast distributions}

\item{kappa}{scale factor}

\item{alpha}{normalized loss when outcome y exceeds forecast x}

\item{w}{numeric vector with cost per unit resource allocated to each
coordinate}

\item{dg}{numeric constant(s) or function(s) to calculate the derivative of the
gpl function \code{g} for each coordinate}

\item{...}{ignored}
}
\value{
a function with argument x that calculates the marginal expected benefit
}
\description{
Create function to calculate the marginal expected benefit of allocating an
additional unit of resources to a given target.
}
