% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{oracle_allocate}
\alias{oracle_allocate}
\title{Allocate according to an oracle's knowledge of outcome y}
\usage{
oracle_allocate(df = NULL, y, K, ...)
}
\arguments{
\item{df}{data frame containing columns for other list arguments which are used only when
those arguments (e.g. \code{F}) are not passed directly}

\item{K}{vector of constraints on total provision}
}
\value{
see \code{allocate}
}
\description{
Allocate according to an oracle's knowledge of outcome y
}
