% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{oracle_alloscore}
\alias{oracle_alloscore}
\title{Score according to an oracle's knowledge of outcome y}
\usage{
oracle_alloscore(df = NULL, y, K, kappa = 1, alpha = 1, g = "x", ...)
}
\arguments{
\item{df}{an allocated data frame}

\item{y}{numeric observed data value}

\item{K}{vector of constraints on total provision (cannot be supplied via df)}

\item{kappa}{scale factor}

\item{alpha}{normalized loss when outcome y exceeds forecast x}

\item{g}{list of strings describing increment functions which will be used
to form gpl functions for each coordinate; should be written using \code{x}, e.g. "log(x)";
will be differentiated using \code{stats::D} so don't try anything fancy; defaults to NULL
deferring to \code{dg}}
}
\value{
see \code{alloscore}
}
\description{
Score according to an oracle's knowledge of outcome y
}
