% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pdqr_factory}
\alias{pdqr_factory}
\title{Produce a pdqr function with set parameters}
\usage{
pdqr_factory(..., dist, type, trans = NULL, trans_inv = NULL, transpars = NULL)
}
\arguments{
\item{...}{list including named parameter values to be partialed out}

\item{dist}{distribution root name or \code{distfromq}}

\item{type}{p, d, q, or r}

\item{transpars}{}
}
\value{
function of type p, d, q, or r
}
\description{
Produce a pdqr function with set parameters
}
