% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stdize_met_params}
\alias{stdize_met_params}
\title{Convert meteorologist parameters to kappa, alpha}
\usage{
stdize_met_params(C, L)
}
\arguments{
\item{C}{marginal cost per unit of recommended protection}

\item{L}{marginal loss due to under-provision of needed resources}
}
\value{
kappa and alpha as list
}
\description{
Convert meteorologist parameters to kappa, alpha
}
